+++ CHANGELOG.rst
@@ -7,6 +7,13 @@ Change Log
 ----------
 
 
+3.1.0
+=====
+
+* Better error diagnostics for S3FS-mounted files that are glaciated
+  if the ``CGAP_S3FS_UPLOAD_BUCKETS`` and ``CGAP_S3FS_UPLOAD_DIR`` environment variables are set.
+
+
 3.0.0
 =====